Backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
faiss-cpu
sentence-transformers
PyYAML
PyPDF2
pdfplumber
plotly
torch
pillow
scikit-learn
ollama
python-multipart
jinja2
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy>=1.4
asyncpg
alembic
pyotp
python-dotenv
email-validator
psycopg2-binary
aiosmtplib
redis>=4.5.1
slowapi>=0.1.9

# EHR Polling
apscheduler>=3.10.0
pyjwt>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-cov>=4.1.0
faker>=19.0.0
